webui: move react components to a new `components` folders
I would like to start keeping react (components code) seperate from the
APIs, as I plan to create wrapper functions for the dbus APIs.

Following the suggestion from https://reactjs.org/docs/faq-structure.html#grouping-by-file-type
