+++ ui/webui/src/components/installation/InstallationProgress.scss
