+++ ui/webui/src/components/installation/ReviewConfiguration.jsx
