+++ ui/webui/src/index.js
@@ -21,7 +21,7 @@ import "./lib/patternfly/patternfly-4-cockpit.scss";
 
 import React from "react";
 import ReactDOM from "react-dom";
-import { Application } from "./app.jsx";
+import { Application } from "./components/app.jsx";
 /*
  * PF4 overrides need to come after the JSX components imports because
  * these are importing CSS stylesheets that we are overriding
@@ -30,7 +30,7 @@ import { Application } from "./app.jsx";
  * the overrides will be correctly in the end of our stylesheet.
  */
 import "./lib/patternfly/patternfly-4-overrides.scss";
-import "./app.scss";
+import "./components/app.scss";
 
 document.addEventListener("DOMContentLoaded", function () {
     ReactDOM.render(React.createElement(Application, {}), document.getElementById("app"));
